src/gateway_api_service/requirements.txt
Flask[async]==3.0.0
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
pydantic>=2.0.0
typing
requests
python-dotenv
pyopenssl